+++ src/Stormancer.Plugins/Blobs/Stormancer.Server.Plugins.BlobStorage/Changelog.rst
@@ -16,6 +16,8 @@ Added
 - Added get blob content API.
 - Added Staging & commit blocks API to upload big blobs.
 - Added IBlobStorageEventHandler extensibility point.
+- stageBlock and commitBlocks now return their success status
+
 
 0.1.0.6
 ----------